+++ packages/test-utils/src/test-project/output/file/GetFileMetadataRequestCommand.ts
@@ -57,7 +57,7 @@ export class GetFileMetadataRequestCommand
   public constructor(input: GetFileMetadataRequestCommandInput) {
     super();
 
-    this.header = { ...input.header, ...defaultHeader };
+    this.header = { ...defaultHeader, ...input.header };
 
     this.param = input.param;
   }